Revert "Rename integral package to integration"
This reverts commit c316acf8
